% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adpt_seeds_match.R
\name{graph_match_adpt_seeds}
\alias{graph_match_adpt_seeds}
\alias{graph_match_soft_seeds}
\title{Graph matching with adpative seeds or soft seeds}
\usage{
graph_match_adpt_seeds(A, B, seeds, non_seed_core = !seeds, select_seeds,
  start = "convex")

graph_match_soft_seeds(A, B, seeds, non_seed_core = !seeds, select_seeds,
  start = "bari")
}
\arguments{
\item{A}{A matrix or an igraph object. Adjacency matrix of \eqn{G_1}.}

\item{B}{A matrix or an igraph object. Adjacency matrix of \eqn{G_2}.}

\item{seeds}{A logical vector. \code{TRUE} indicates the corresponding
vertex is a seed.}

\item{non_seed_core}{A logical vector. \code{TRUE} indicates the corresponding
vertex is a core vertex and is not a seed. Note in the case without junk vertices,
\code{non_seed_core=!seeds}.}

\item{select_seeds, }{A vector. A vector, a matrix or a data frame. If there is no error in soft
seeds, input can be a vector of soft seed indices in \eqn{G_1}. Or if there is error in soft
seeds, input in the form of a matrix or a data frame, with the first column being the
indices of \eqn{G_1} and the second column being the corresponding indices of \eqn{G_2}. Note
that if there are seeds in graphs, seeds should be put before non-seeds.}

\item{start}{A matrix or a character. Any \code{nns-by-nns} matrix or
character value like "bari", "convex" or "rds" to initialize the starting matrix.}
}
\value{
Returns a list of graph matching results, including core matching error, ultimate matching
error in selected seeds, and value for the objective function.
}
\description{
Match two given graphs with a set of selected best matched nodes as additional seeds.
}
\examples{
cgnp_pair <- sample_correlated_gnp_pair(n = 50, p =  0.5, rho = 0.3)
g1 <- cgnp_pair$graph1
g2 <- cgnp_pair$graph2
seeds <- 1:50 <= 10
match <- graph_match_FW(g1, g2, seeds, start = "bari")
select_seeds <- select_seeds(g1, g2, "row_cor", nseeds=5, x=!seeds, match_corr=match$corr)
match_adpt <- graph_match_adpt_seeds(g1, g2, seeds, select_seeds=select_seeds)
match_adpt

match_soft <- graph_match_soft_seeds(g1, g2, seeds, select_seeds=select_seeds)
match_soft
}
