% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splr_matrix_class.R
\name{as.character.splrMatrix}
\alias{as.character.splrMatrix}
\title{splr "Matrix" as character}
\usage{
\method{as.character}{splrMatrix}(x, ...)
}
\arguments{
\item{x}{splrMatrix}
}
\value{
character output of splr matrix
}
\description{
splr "Matrix" as character
}
\keyword{internal}
