% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_matches.R
\name{best_matches}
\alias{best_matches}
\title{Choose best matches}
\usage{
best_matches(A, B, measure, num, x, match_corr)
}
\arguments{
\item{A}{A matrix or an igraph object. Adjacency matrix of \eqn{G_1}.}

\item{B}{A matrix or an igraph object. Adjacency matrix of \eqn{G_2}.}

\item{measure}{A character. Measure for computing goodness of matching.}

\item{num}{An integer. Number of pairs of best matched vertices needed.}

\item{x}{A vector of logical. \code{TRUE} indicates the corresponding vertex is of interest
in finding best matched vertices. Length of vector should be the number of vertice of graphs.}

\item{match_corr}{A vector. Graph matching correspondence between \eqn{G_1} and \eqn{G_2}.}
}
\value{
\code{best_matches} returns a data frame with the indices of best matched vertices
in \eqn{G_1} named \code{A_best} and the indices of best matched vertices in \eqn{G_2} named
\code{B_best}.
}
\description{
Find a set of vertices pairs in the order of goodness of matching according to a
specified measure.
}
\examples{
cgnp_pair <- sample_correlated_gnp_pair(n = 50, corr =  0.3, p =  0.5)
g1 <- cgnp_pair$graph1
g2 <- cgnp_pair$graph2
seeds <- 1:50 <= 10
nonseeds <- !seeds
match <- graph_match_FW(g1, g2, seeds)

# Application: select best matched seeds from non seeds as new seeds, and do the
# graph matching iteratively to get higher matching accuracy
best_matches(g1, g2, "row_perm_stat", num = 5, x = nonseeds, match$corr)


}
