% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_soft_seeds.R
\name{check_soft_seeds}
\alias{check_soft_seeds}
\title{Check initialization for soft seeding}
\usage{
check_soft_seeds(start, nns, ns = 0, soft_seeds, A = NULL, B = NULL,
  seeds = NULL, non_seed_core = NULL)
}
\arguments{
\item{start}{A matrix or a character. Any \code{nns-by-nns} matrix or character value
like "bari", "convex" or "rds" to initialize the starting matrix.}

\item{nns}{An integer. Number of non-seeds.}

\item{ns}{An integer. Number of seeds.}

\item{soft_seeds}{A vector, a matrix or a data frame. If there is no error in soft
seeds, input can be a vector of soft seed indices in \eqn{G_1}. Or if there is error in soft
seeds, input in the form of a matrix or a data frame, with the first column being the
indices of \eqn{G_1} and the second column being the corresponding indices of \eqn{G_2}. Note
that if there are seeds in graphs, seeds should be put before non-seeds.}

\item{A}{A matrix or an igraph object. Adjacency matrix of \eqn{G_1}. Needed only when start is
convex.}

\item{B}{A matrix or an igraph object. Adjacency matrix of \eqn{G_2}. Needed only when start is
convex.}

\item{seeds}{A logical vector. \code{TRUE} indicates the corresponding vertex is a seed. Needed
only when start is convex.}

\item{non_seed_core}{A logical vector. \code{TRUE} indicates the corresponding vertex is a
non_seed_core vertex. Needed only when start is convex.}
}
\value{
\code{check_soft_seeds} returns a \code{nns-by-nns} matrix with 1's corresponding to the
adaptive seeds and being bari-centered at other places.
}
\description{
Initialize the start matrix for soft seeding graph matching iteration.
}
\examples{
ss <- matrix(c(2,4,2,3), nrow = 2)
check_soft_seeds("bari", 5, 3, c(5,7,8))
check_soft_seeds(start = "rds", nns = 5, soft_seeds = ss)

cgnp_pair <- sample_correlated_gnp_pair(n = 50, rho =  0.3, p =  0.5)
g1 <- cgnp_pair$graph1
g2 <- cgnp_pair$graph2
seeds <- 1:50 <= 10
ns_core <- 1:50 <= 50
check_soft_seeds(start = "convex", nns = 5, soft_seeds = ss, A = g1, B = g2, seeds = seeds, non_seed_core = ns_core)

}
