% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Enron}
\alias{Enron}
\title{Email communication networks of Enron Corporation}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(Enron)
}
\description{
The Enron network data consists of email messages between 184 employees
of the Enron Corporation where each graph represents one week of emails
and each edge indicates whether there is email sent from one employee to
the other.
Two networks are unweighted and directed with self-loops.
There are 488 and 482 edges in two networks respectively and the empirical
Pearson's correlation between two graphs is 0.85.
Two email communication networks for two different weeks are stored in a
list in the form of igraph objects.
}
\examples{
data(Enron)
g1 <- Enron[[1]]
g2 <- Enron[[2]]

}
\references{
Originally released by William Cohen at CMU. \href{http://www.cs.cmu.edu/~enron/}{More details}
on the origins and research uses of the dataset.
}
\keyword{datasets}
