% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_convex.R, R/graph_match_methods.R
\name{graph_match_convex}
\alias{graph_match_convex}
\alias{graph_match_FW}
\title{Graph Match - Frank Wolf}
\usage{
graph_match_convex(
  A,
  B,
  seeds = NULL,
  start = "bari",
  max_iter = 100,
  similarity = NULL,
  tol = 1e-05,
  lap_method = NULL
)

graph_match_FW(
  A,
  B,
  seeds = NULL,
  start = "bari",
  max_iter = 20,
  similarity = NULL,
  lap_method = NULL
)
}
\arguments{
\item{A}{A matrix, igraph object, or list of either.}

\item{B}{A matrix, igraph object, or list of either.}

\item{seeds}{A vector of integers or logicals, a matrix or a data frame. If
the seed pairs have the same indices in both graphs then seeds can be a
vector. If not, seeds must be  a matrix
or a data frame, with the first column being the indices of \eqn{G_1} and
the second column being the corresponding indices of \eqn{G_2}.}

\item{start}{A matrix or a character. Any \code{nns-by-nns} matrix or
character value like "bari" or "convex" to initialize the starting matrix.}

\item{max_iter}{A number. Maximum number of replacing matches equals to
max_iter times number of total vertices of \eqn{G_1}.}

\item{similarity}{A matrix. An \code{n-by-n} matrix containing vertex similaities.}

\item{tol}{A number. Tolerance of edge disagreements.}

\item{lap_method}{Choice for lap method.}

\item{alpha}{A number betwen 0 and 1. Bigger alpha means putting more importance
on the information in network topology over other information such as
similarity scores}

\item{method}{A character. Choice of method to extract mapping from score matrix,
including greedy method and the Hungarian algorithm.}
}
\value{
\code{graph_match_convex} returns a list of graph matching results,
  including the graph matching formula, a data frame containing the matching 
  correspondence between \eqn{G_1} and \eqn{G_2} named \code{corr_A} and 
  \code{corr_B} and the number of seeds.

\code{graph_match_FW} returns a list of graph matching results,
  including the graph matching formula, a data frame containing the matching 
  correspondence between \eqn{G_1} and \eqn{G_2} named \code{corr_A} and 
  \code{corr_B} and the number of seeds.
}
\description{
Match two given graphs, returns a list of graph matching
  results, including matching correspondence vector of \eqn{G_2} with respect
  to \eqn{G_1}, doubly stochastic matrix and permutation matrix.
}
\examples{
seeds <- 1:10 <= 3
graph_match_convex(g1, g2, seeds)

hard_seeds <- matrix(c(4,6,5,4),2)
seeds <- rbind(as.matrix(check_seeds(seeds, 10)$seeds),hard_seeds)
\dontrun{
# for some reason this fails in check
graph_match_convex(g1, g2, seeds)
}

cgnp_pair <- sample_correlated_gnp_pair(n = 10, corr =  0.3, p =  0.5)
g1 <- cgnp_pair$graph1
g2 <- cgnp_pair$graph2
# match G_1 & G_2 with no seeds
graph_match_FW(g1, g2)

# match G_1 & G_2 with some known node pairs as seeds
seeds <- 1:10 <= 3
graph_match_FW(g1, g2, seeds, start = "bari")

# match G_1 & G_2 with some incorrect seeds
hard_seeds <- matrix(c(4,6,5,4),2)
seeds <- rbind(as.matrix(check_seeds(seeds, nv = 10)$seeds),hard_seeds)
graph_match_FW(g1, g2, seeds, start = "convex")

 gp_list <- replicate(3, sample_correlated_gnp_pair(100, .3, .5), simplify = FALSE)
 A <- lapply(gp_list, function(gp)gp[[1]])
 B <- lapply(gp_list, function(gp)gp[[2]])
 match <- graph_match_FW(A, B, seeds = 1:10, start = "bari", max_iter = 20)
 match$corr

}
