% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_match_multi.R
\name{graph_match_FW_multi}
\alias{graph_match_FW_multi}
\title{Multiple Graph Match Methods}
\usage{
graph_match_FW_multi(A, B, seeds = NULL, start = "bari", max_iter = 20,
  similarity = NULL, usejv = TRUE)
}
\arguments{
\item{A}{A list of graphs (or igraph objects)}

\item{B}{A list of graphs (or igraph objects)}

\item{seeds}{A vector of integers or logicals, a matrix or a data frame. If there is no error in seeds input can be
a vector of seed indices in \eqn{G_1}. Or if there exists error in seeds, input in the form of a
matrix or a data frame, with the first column being the indices of \eqn{G_1} and the second
column being the corresponding indices of \eqn{G_2}.}

\item{start}{A matrix or a character. Any \code{nns-by-nns} matrix or
character value like "bari" or "convex" to initialize the starting matrix.}

\item{max_iter}{An integer. Maximum iteration time.}

\item{tol}{A number. Tolerance of edge disagreements.}

\item{r}{A number. Threshold of neighboring pair scores.}
}
\value{
\code{graph_match_FW_multi} returns a list of graph matching results,
including match correspondence vector of \eqn{G_2} with respect to \eqn{G_1}
named \code{corr}, doubly stochastic matrix named \code{D}, permutation
matrix named \code{P} based on Frank-Wolfe methodology and iteration time of
the algorithm named \code{iter}.
}
\description{
Match two lists of graphs, returns a list of graph matching results,
including matching correspondence vector of \eqn{G_2} with respect to \eqn{G_1},
doubly stochastic matrix and permutation matrix.
}
\examples{
 gp_list <- replicate(3, sample_correlated_gnp_pair(100, .3, .5), simplify = FALSE)
 A <- lapply(gp_list, function(gp)gp[[1]])
 B <- lapply(gp_list, function(gp)gp[[2]])
 match <- graph_match_FW_multi(A, B, seeds = 1:10, start = "bari", max_iter = 20)
 match$corr

}
