% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_match_multi.R
\name{graph_match_mutual_multi}
\alias{graph_match_mutual_multi}
\title{export}
\usage{
graph_match_mutual_multi(A, B, start = NULL, similarity = NULL,
  alpha = 0.2, max_iter = 50)
}
\description{
export
}
