% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_match_class.R
\name{as.data.frame,graphMatch-method}
\alias{as.data.frame,graphMatch-method}
\alias{show,graphMatch-method}
\alias{print,graphMatch-method}
\alias{[,graphMatch,missing,missing,missing-method}
\alias{dim,graphMatch-method}
\alias{length,graphMatch-method}
\alias{t,graphMatch-method}
\alias{rev,graphMatch-method}
\alias{[,graphMatch,ANY,ANY,ANY-method}
\alias{str,graphMatch-method}
\alias{$,graphMatch-method}
\title{Methods for the graphMatch class}
\usage{
\S4method{as.data.frame}{graphMatch}(x)

\S4method{show}{graphMatch}(object)

\S4method{print}{graphMatch}(x)

\S4method{[}{graphMatch,missing,missing,missing}(x, i = NULL, j = NULL, drop = NULL)

\S4method{dim}{graphMatch}(x)

\S4method{length}{graphMatch}(x)

\S4method{t}{graphMatch}(x)

\S4method{rev}{graphMatch}(x)

\S4method{[}{graphMatch,ANY,ANY,ANY}(x, i = NULL, j = NULL, drop = NULL)

\S4method{str}{graphMatch}(object)

\S4method{$}{graphMatch}(x, name)
}
\arguments{
\item{x}{graphMatch object}

\item{object}{graphMatch object}

\item{i}{row index for the correspondence data.frame}

\item{j}{col index for the correspondence data.frame}

\item{drop}{ignored}

\item{name}{name of element in the list}
}
\value{
dim returns a vector of length two 
indicating the number of vertices in each original graph.
length returns the number of found vertex-pair matches.
m[i,j] will index the 2xlength data.frame of vertex-pair matches.
This is true for any i,j unless both are missing.
In that case, m[] returns a sparse matrix of dimension dim(m)
where m[][i,j] is 0 unless m matches node i with node j.
(Note this is not guaranteed to be a permutation matrix unless
dim(m)[1] = dim(m)[2] = length(m).
}
\description{
These methods provide functionality to plot,
 view, inspect, and convert graphMatch objects.
}
\details{
Methods for the graphmatch
}
\seealso{
graphMatch_operators

graphMatch_constructor
}
