% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_match_class.R
\name{\%*\%,graphMatch,ANY-method}
\alias{\%*\%,graphMatch,ANY-method}
\title{Operator methods for graphMatch objects}
\usage{
\S4method{\%*\%}{graphMatch,ANY}(x, y)
}
\description{
Methods to use graphMatch objects as operators on 
 igraph and matrix-like objects.
}
