% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_match_class.R
\name{summary,graphMatch-method}
\alias{summary,graphMatch-method}
\title{Summary methods for graphMatch objects}
\usage{
\S4method{summary}{graphMatch}(object, A = NULL, B = NULL, true_label = NULL, directed = NULL, ...)
}
\description{
Summary methods for graphMatch objects
}
