% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iGraphMatch-package.R
\docType{package}
\name{iGraphMatch-package}
\alias{iGraphMatch}
\alias{iGraphMatch-package}
\title{iGraphMatch: Tools for approximately solving graph matching problems}
\description{
iGraphMatch is an R package for graph
    matching. The package works for both igraph objects and matrix
    objects. You provide the adjacency matrices of two graphs and some
    other information you might know, choose the graph matching method,
    and it returns the graph matching results. iGraphMatch also provides a
    bunch of useful functions you might need related to graph
    matching.
}
\author{
\strong{Maintainer}: Zihuan Qiao \email{zhqiao@bu.edu}

Authors:
\itemize{
  \item Daniel Sussman \email{sussman@bu.edu}
}

}
\keyword{internal}
