% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splr_matrix_class.R
\name{innerproduct}
\alias{innerproduct}
\alias{innerproduct,splrMatrix,splrMatrix-method}
\alias{innerproduct,splrMatrix,Matrix-method}
\alias{innerproduct,Matrix,splrMatrix-method}
\alias{innerproduct,matrix_list,matrix_list-method}
\title{Matrix inner products}
\usage{
innerproduct(x, y)

\S4method{innerproduct}{splrMatrix,splrMatrix}(x, y)

\S4method{innerproduct}{splrMatrix,Matrix}(x, y)

\S4method{innerproduct}{Matrix,splrMatrix}(x, y)

\S4method{innerproduct}{matrix_list,matrix_list}(x, y)
}
\arguments{
\item{x}{matrix like object}

\item{y}{matrix like object}
}
\value{
inner product <x, y> = sum over all elements i,j of x_ij * y_ij.
}
\description{
Matrix inner products
}
\details{
For a matrix_list object, sums over all layers/list-elements as well.
}
\keyword{internal}
