% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlapjv.R
\name{lapmod}
\alias{lapmod}
\title{Solves the linear assignment problem using the LAPMOD algorithm}
\usage{
lapmod(cost, maximize = FALSE)
}
\arguments{
\item{cost}{A non-negative 'CsparseMatrix' object from the 'Matrix' package}

\item{maximize}{If FALSE (default) then costs are minimized and if TRUE the
costs are maximized}
}
\value{
The assignment of rows to columns as an integer vector
}
\description{
Find a set of vertices pairs in the order of goodness of matching according to a
specified measure.
}
\details{
The 'C++' code for this method is modified from code in the
 \href{https://github.com/Bram94/lapjv}{python lapjv} package.
}
