% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlapjv.R
\name{lapmod_index}
\alias{lapmod_index}
\title{Solves the linear assignment problem using the LAPMOD algorithm}
\usage{
lapmod_index(n, cc, ii, kk, maximize = FALSE)
}
\arguments{
\item{n}{number of rows in the cost matrix}

\item{cc}{vector of all finite elements of the assignement cost matri}

\item{ii}{vector of indices of the zero indexed  row starts in cc. The following must hold
ii[1] = 0 and ii[n+2] = length(cc).}

\item{kk}{0-based column numbers for each finite cost in the matrix, 
i.e., kk must be in 0:(nrow(.)-1).}

\item{maximize}{If FALSE (default) then costs are minimized and if TRUE the 
costs are maximized}
}
\value{
The assignment of rows to columns as an integer vector
}
\description{
Find a set of vertices pairs in the order of goodness of matching according to a
specified measure.
}
