% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_report.R
\name{match_report}
\alias{match_report}
\title{Matching performance summary}
\usage{
match_report(object, A = A, B = B, label = TRUE, ...)
}
\arguments{
\item{object}{A list. calls the matching result of applying a specific graph matching
algorithm.}

\item{label}{A logical. TRUE if the true correspondence between two graphs is known, such
as the simulated data.}
}
\value{
\code{match_report} returns a list of matching performance evaluation metrics
including number of matches, true matches, common edges, common non-edges, edge correctness
which is the fraction of common edges over number of edges in the first graph, and the
objective value ||A-PBP^T||_F. The function also returns the matching result including the
correspondence and the permutation matrix.
}
\description{
Get a summary of the matching result and measures of the matching performance
based on several evaluation metrics associated with nodes and edges of two graphs.
}
\examples{
graphs <- sample_correlated_gnp_pair(10, .5, .3)
A <- graphs$graph1
B <- graphs$graph2
res <- graph_match_percolation(A, B, 1:4)
match_report(res, A, B)

}
