% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_list.R
\docType{class}
\name{matrix_list-class}
\alias{matrix_list-class}
\alias{matrix_list}
\alias{\%*\%,matrix_list,matrix_list-method}
\alias{t,matrix_list-method}
\alias{dim,matrix_list-method}
\alias{[,matrix_list,ANY,ANY,ANY-method}
\alias{[,matrix_list,ANY,missing,ANY-method}
\alias{[,matrix_list,missing,ANY,ANY-method}
\alias{[,matrix_list,missing,missing,ANY-method}
\alias{\%*\%,matrix_list,ANY-method}
\alias{\%*\%,ANY,matrix_list-method}
\alias{sum,matrix_list,logical-method}
\alias{^,matrix_list,ANY-method}
\alias{-,matrix_list,matrix_list-method}
\alias{+,matrix_list,matrix_list-method}
\alias{*,matrix_list,matrix_list-method}
\alias{/,matrix_list,matrix_list-method}
\alias{-,matrix_list,ANY-method}
\alias{+,matrix_list,ANY-method}
\alias{*,matrix_list,ANY-method}
\alias{/,matrix_list,ANY-method}
\alias{-,ANY,matrix_list-method}
\alias{+,ANY,matrix_list-method}
\alias{*,ANY,matrix_list-method}
\alias{/,ANY,matrix_list-method}
\alias{-,matrix_list,missing-method}
\alias{names<-,matrix_list-method}
\alias{tcrossprod,matrix_list,matrix_list-method}
\alias{tcrossprod,ANY,matrix_list-method}
\alias{tcrossprod,matrix_list,ANY-method}
\title{Lists of Matrices}
\usage{
matrix_list(ml)

\S4method{\%*\%}{matrix_list,matrix_list}(x, y)

\S4method{t}{matrix_list}(x)

\S4method{dim}{matrix_list}(x)

\S4method{[}{matrix_list,ANY,ANY,ANY}(x, i = 1:nrow(x[[1]]), j = 1:ncol(x[[1]]), drop = FALSE)

\S4method{[}{matrix_list,ANY,missing,ANY}(x, i, drop = FALSE)

\S4method{[}{matrix_list,missing,ANY,ANY}(x, j, drop = FALSE)

\S4method{[}{matrix_list,missing,missing,ANY}(x, drop = FALSE)

\S4method{\%*\%}{matrix_list,ANY}(x, y)

\S4method{\%*\%}{ANY,matrix_list}(x, y)

\S4method{sum}{matrix_list,logical}(x, na.rm = FALSE)

\S4method{^}{matrix_list,ANY}(e1, e2)

\S4method{-}{matrix_list,matrix_list}(e1, e2)

\S4method{+}{matrix_list,matrix_list}(e1, e2)

\S4method{*}{matrix_list,matrix_list}(e1, e2)

\S4method{/}{matrix_list,matrix_list}(e1, e2)

\S4method{-}{matrix_list,ANY}(e1, e2)

\S4method{+}{matrix_list,ANY}(e1, e2)

\S4method{*}{matrix_list,ANY}(e1, e2)

\S4method{/}{matrix_list,ANY}(e1, e2)

\S4method{-}{ANY,matrix_list}(e1, e2)

\S4method{+}{ANY,matrix_list}(e1, e2)

\S4method{*}{ANY,matrix_list}(e1, e2)

\S4method{/}{ANY,matrix_list}(e1, e2)

\S4method{-}{matrix_list,missing}(e1, e2)

\S4method{names}{matrix_list}(x) <- value

\S4method{tcrossprod}{matrix_list,matrix_list}(x, y)

\S4method{tcrossprod}{ANY,matrix_list}(x, y)

\S4method{tcrossprod}{matrix_list,ANY}(x, y)
}
\arguments{
\item{ml}{A list of matrices}

\item{x}{As in Matrix}

\item{y}{As in Matrix}

\item{i}{As in Matrix}

\item{j}{As in Matrix}

\item{drop}{As in Matrix}

\item{na.rm}{As in Matrix}

\item{e1}{As in Matrix}

\item{e2}{As in Matrix}
}
\description{
Basically a list with matrix components that are all 
the same dimension. Mostly for internal igraphmatch use.
It can do various things like arithmetic and
indexing, all of which are done component-wise.
}
\keyword{internal}
