% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center_graph.R
\name{pad}
\alias{pad}
\title{Pad a matrix object with extra rows/columns of 0s.}
\usage{
pad(m, nr, nc = nr)
}
\arguments{
\item{m}{matrix}

\item{nr}{number of rows to add}

\item{nc}{number of columns to add. (default = nr)}
}
\value{
m padded with nr rows and nc columns of zeros.
}
\description{
Attempts are made to make this padding efficient
by employing sparse graphs
}
