% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_func.R
\name{row_cor}
\alias{row_cor}
\title{Measure functions}
\usage{
row_cor(g1, g2)
}
\arguments{
\item{g1}{A matrix or an 'igraph' object. Adjacency matrix of \eqn{G_1}.}

\item{g2}{A matrix or an 'igraph' object. Adjacency matrix of \eqn{G_2} after adjusting rows and columns
according to the correlation of matching between two graphs.}

\item{exact}{A logical. If \code{g1} and \code{g2} are binary, then set \code{exact=TRUE},
if \code{g1} and \code{g2} are weighted graphs, then set \code{exact=FALSE}.}
}
\description{
Measures for computing the goodness of matching for each vertex.
}
