% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_perm.R
\name{rperm}
\alias{rperm}
\title{Sample random permutation matrix}
\usage{
rperm(n)
}
\arguments{
\item{n}{An integer. Dimension of the permutation matrix.}
}
\value{
\code{rperm} returns an n-by-n permutation matrix.
}
\description{
Sample an n-by-n random permutation matrix.
}
\examples{
rperm(3)
}
