% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_seeds.R
\name{select_seeds}
\alias{select_seeds}
\title{Select adaptive seeds}
\usage{
select_seeds(A, B, measure, nseeds, x, match_corr)
}
\arguments{
\item{A}{A matrix or an igraph object. Adjacency matrix of \eqn{G_1}.}

\item{B}{A matrix or an igraph object. Adjacency matrix of \eqn{G_2}.}

\item{measure}{A character. Measure for computing goodness of matching.}

\item{nseeds}{An integer. Number of adaptive seeds needed.}

\item{x}{A vector of logical. \code{TRUE} indicates the corresponding vertex is of interest
in finding adaptive seeds. Length of vector should be the number of vertice of graphs.}

\item{match_corr}{A vector. Graph matching correspondence between \eqn{G_1} and \eqn{G_2}.}
}
\value{
\code{select_seeds} returns a data frame with the indice of seeds in \eqn{G_1} named
\code{seed_A} and the indice of seeds in \eqn{G_2} named \code{seed_B}.
}
\description{
Find a set of vertex pairs that are best matched be to adaptive seeds.
}
\examples{
cgnp_pair <- sample_correlated_gnp_pair(n = 50, corr =  0.3, p =  0.5)
g1 <- cgnp_pair$graph1
g2 <- cgnp_pair$graph2
seeds <- 1:50 <= 10
nonseeds <- !seeds
match <- graph_match_FW(g1, g2, seeds)

# select adaptive seeds from non seeds
select_seeds(g1, g2, "row_perm_stat", nseeds = 5, x = nonseeds, match$corr)

}
