% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splr_matrix_class.R
\name{splr}
\alias{splr}
\title{splrMatrix constructor}
\usage{
splr(x, a = NULL, b = NULL, rank = NULL, factorize = FALSE,
  tol = 1e-05, dimnames = list(NULL, NULL))
}
\description{
Construct a sparse plus low rank matrix
}
