% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splr_matrix_class.R
\name{splr_to_sparse}
\alias{splr_to_sparse}
\title{Convert splr 'Matrix' to Sparse}
\usage{
splr_to_sparse(data)
}
\arguments{
\item{data}{splrMatrix}
}
\value{
sparse Matrix equal to x + a %*% t(b)

See \code{\link[Matrix]{Matrix}}.
}
\description{
Convert splr 'Matrix' to Sparse
}
