% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_match_class.R
\name{as}
\alias{as}
\alias{as.data.frame,graphMatch-method}
\alias{show,graphMatch-method}
\alias{print,graphMatch-method}
\alias{[,graphMatch,missing,missing,missing-method}
\alias{dim,graphMatch-method}
\alias{length,graphMatch-method}
\alias{[,graphMatch,ANY,ANY,ANY-method}
\alias{str,graphMatch-method}
\title{Methods for the graphMatch class}
\usage{
\S4method{as.data.frame}{graphMatch}(x)

\S4method{show}{graphMatch}(object)

\S4method{print}{graphMatch}(x)

\S4method{[}{graphMatch,missing,missing,missing}(x, i = NULL, j = NULL, drop = NULL)

\S4method{dim}{graphMatch}(x)

\S4method{length}{graphMatch}(x)

\S4method{[}{graphMatch,ANY,ANY,ANY}(x, i = NULL, j = NULL, drop = NULL)

\S4method{str}{graphMatch}(object)
}
\description{
These methods provide functionality to plot,
 view, inspect, and convert graphMatch objects.
}
