% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iGraphMatch-package.R
\docType{package}
\name{iGraphMatch-package}
\alias{iGraphMatch}
\alias{iGraphMatch-package}
\title{iGraphMatch: Tools Graph Matching}
\description{
Versatile tools and data for graph matching analysis with various forms of prior information
    that supports working with igraph objects, matrix objects, or lists of either.
}
\author{
\strong{Maintainer}: Daniel Sussman \email{sussman@bu.edu}

Authors:
\itemize{
  \item Zihuan Qiao \email{zhqiao@bu.edu}
}

Other contributors:
\itemize{
  \item Joshua Agterberg [contributor]
  \item Lujia Wang [contributor]
  \item Vince Lyzinski [contributor]
}

}
\keyword{internal}
