% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splr_matrix_class.R
\name{innerproduct}
\alias{innerproduct}
\alias{innerproduct,splrMatrix,splrMatrix-method}
\alias{innerproduct,splrMatrix,Matrix-method}
\alias{innerproduct,Matrix,splrMatrix-method}
\alias{innerproduct,matrix_list,matrix_list-method}
\title{Matrix inner products}
\usage{
innerproduct(x, y)

\S4method{innerproduct}{splrMatrix,splrMatrix}(x, y)

\S4method{innerproduct}{splrMatrix,Matrix}(x, y)

\S4method{innerproduct}{Matrix,splrMatrix}(x, y)

\S4method{innerproduct}{matrix_list,matrix_list}(x, y)
}
\arguments{
\item{x}{matrix like object}

\item{y}{matrix like object}
}
\value{
inner product <x, y>
}
\description{
Matrix inner products
}
