% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_report.R
\name{match_report}
\alias{match_report}
\title{Matching report}
\usage{
match_report(A, B, corr, seeds = NULL, label = TRUE)
}
\arguments{
\item{A}{A matrix or an igraph object. Adjacency matrix of \eqn{G_1}.}

\item{B}{A matrix or an igraph object. Adjacency matrix of \eqn{G_2}.}

\item{corr}{A matrix or a data frame. Matching correspondence with the first
and second columns correspond to indices in \eqn{G_1} and \eqn{G_2} respectively.}

\item{seeds}{A vector of integers or logicals, a matrix or a data frame. If
there is no error in seeds input can be a vector of seed indices in
\eqn{G_1}. Or if there exists error in seeds, input in the form of a matrix
or a data frame, with the first column being the indices of \eqn{G_1} and
the second column being the corresponding indices of \eqn{G_2}.}

\item{label}{A logical. TRUE if the true correspondence is known.}
}
\value{
\code{match_report} returns a list of matching performance evaluation metrics
including number of matches, true matches, common edges, common non-edges, edge correctness
which is the fraction of common edges over number of edges in the first graph, and the
objective value ||A-PBP^T||_F.
}
\description{
Get a report of matching performance based on several evaluation metrics
 according to the mapping correspondence.
}
\examples{
graphs <- sample_correlated_gnp_pair(10, .5, .3)
A <- graphs$graph1
B <- graphs$graph2
corr <- graph_match_percolation(A, B, 1:4)$corr
match_report(A, B, corr, seeds = 1:4)

}
