% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_report.R
\name{matched_adjs}
\alias{matched_adjs}
\title{document}
\usage{
matched_adjs(match, A, B)
}
\arguments{
\item{match}{Result from a a graph matching method.}

\item{A}{A matrix, igraph object, or list of either.
Likely used in the call for creating match.}

\item{B}{A matrix, igraph object, or list of either.
Likely used in the call for creating match.}
}
\description{
Return aligned versions of A and B according to
 a result of match method
}
