% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_list.R
\docType{class}
\name{matrix_list-class}
\alias{matrix_list-class}
\alias{matrix_list}
\alias{\%*\%,matrix_list,matrix_list-method}
\alias{t,matrix_list-method}
\alias{[,matrix_list,ANY,ANY,ANY-method}
\alias{[,matrix_list,ANY,missing,ANY-method}
\alias{[,matrix_list,missing,ANY,ANY-method}
\alias{[,matrix_list,missing,missing,ANY-method}
\alias{\%*\%,matrix_list,ANY-method}
\alias{\%*\%,ANY,matrix_list-method}
\alias{sum,matrix_list,logical-method}
\alias{^,matrix_list,ANY-method}
\alias{-,matrix_list,matrix_list-method}
\alias{+,matrix_list,matrix_list-method}
\alias{*,matrix_list,matrix_list-method}
\alias{/,matrix_list,matrix_list-method}
\alias{-,matrix_list,ANY-method}
\alias{+,matrix_list,ANY-method}
\alias{*,matrix_list,ANY-method}
\alias{/,matrix_list,ANY-method}
\alias{-,ANY,matrix_list-method}
\alias{+,ANY,matrix_list-method}
\alias{*,ANY,matrix_list-method}
\alias{/,ANY,matrix_list-method}
\alias{-,matrix_list,missing-method}
\title{Lists of Matrices}
\usage{
\S4method{\%*\%}{matrix_list,matrix_list}(x, y)

\S4method{t}{matrix_list}(x)

\S4method{[}{matrix_list,ANY,ANY,ANY}(x, i = 1:nrow(x[[1]]), j = 1:ncol(x[[1]]), drop = FALSE)

\S4method{[}{matrix_list,ANY,missing,ANY}(x, i, drop = FALSE)

\S4method{[}{matrix_list,missing,ANY,ANY}(x, j, drop = FALSE)

\S4method{[}{matrix_list,missing,missing,ANY}(x, drop = FALSE)

\S4method{\%*\%}{matrix_list,ANY}(x, y)

\S4method{\%*\%}{ANY,matrix_list}(x, y)

\S4method{sum}{matrix_list,logical}(x, na.rm = FALSE)

\S4method{^}{matrix_list,ANY}(e1, e2)

\S4method{-}{matrix_list,matrix_list}(e1, e2)

\S4method{+}{matrix_list,matrix_list}(e1, e2)

\S4method{*}{matrix_list,matrix_list}(e1, e2)

\S4method{/}{matrix_list,matrix_list}(e1, e2)

\S4method{-}{matrix_list,ANY}(e1, e2)

\S4method{+}{matrix_list,ANY}(e1, e2)

\S4method{*}{matrix_list,ANY}(e1, e2)

\S4method{/}{matrix_list,ANY}(e1, e2)

\S4method{-}{ANY,matrix_list}(e1, e2)

\S4method{+}{ANY,matrix_list}(e1, e2)

\S4method{*}{ANY,matrix_list}(e1, e2)

\S4method{/}{ANY,matrix_list}(e1, e2)

\S4method{-}{matrix_list,missing}(e1, e2)
}
\description{
Basically a list that has matrices that are all 
the same size. It doesn't do much checking though.
It can do various things like arethmatic and
indexing, all of which are done element-wise.
}
