% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_gnp.R
\name{sample_correlated_gnp_pair}
\alias{sample_correlated_gnp_pair}
\alias{sample_correlated_gnp_pair_w_junk}
\title{Sample correlated G(n,p) random graphs}
\usage{
sample_correlated_gnp_pair(n, corr, p, directed = FALSE,
  permutation = NULL)

sample_correlated_gnp_pair_w_junk(n, corr, p, ncore = n)
}
\arguments{
\item{n}{An integer. Number of total vertices for the sampled graphs.}

\item{corr}{A number. The target Pearson correlation between the adjacency matrices
of the generated graphs. It must be in open (0,1) interval.}

\item{p}{A number. Edge probability between two vertices. It must be in open
(0,1) interval.}

\item{directed}{A logical. Whether to generate directed graphs.}

\item{permutation}{A vector of number. A permutation vector that is applied on the
vertices of the first graph, to get the second graph. If \code{NULL}, the vertices
are not permuted.}

\item{ncore}{An integer. Number of core vertices.}
}
\value{
\code{sample_correlated_gnp_pair} returns a list of two igraph object, named
\code{graph1} and \code{graph2}, which are two graphs whose adjacency matrix entries
correlated with \code{rho}.

\code{sample_correlated_gnp_pair_w_junk} returns a list of two igraph object, named
\code{graph1} and \code{graph2}, which are two graphs whose adjacency matrix entries
correlated with \code{rho} and with first ncore vertices being core vertices and the rest being
junk vertices.
}
\description{
Sample a pair of correlated G(n,p) random graphs with correlation between
two graphs being \code{rho} and edge probability being \code{p}.
}
\examples{
sample_correlated_gnp_pair(50, 0.3, 0.5)
sample_correlated_gnp_pair_w_junk(50, 0.3, 0.5, 40)


}
