% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splr_matrix_class.R
\docType{class}
\name{splrMatrix-class}
\alias{splrMatrix-class}
\alias{splr}
\alias{splr,Matrix,Matrix,Matrix-method}
\title{Sparse Plus Low-Rank Matrices}
\usage{
splr(
  x,
  a = NULL,
  b = NULL,
  rank = NULL,
  factorize = FALSE,
  tol = 1e-05,
  dimnames = list(NULL, NULL),
  ...
)

\S4method{splr}{Matrix,Matrix,Matrix}(
  x,
  a = NULL,
  b = NULL,
  rank = NULL,
  factorize = FALSE,
  tol = 1e-05,
  dimnames = list(NULL, NULL),
  ...
)
}
\arguments{
\item{x}{As in Matrix}

\item{a}{As in Matrix}

\item{b}{As in Matrix}

\item{rank}{the estimated rank of the matrix to be factorized.  If \code{rank} is not provided, a guess is made}

\item{factorize}{a safeguard to prevent unnecessary computation (e.g. if the matrix is very large).  If \code{a} 
is provided and \code{factorize = FALSE}, an error will result.}

\item{tol}{the tolerance for the eigenvalues if \code{rank} is not provided}

\item{dimnames}{optional - the list of names for the matrix}

\item{...}{As in Matrix}
}
\description{
An S4 class for efficient computation with sparse plus
low-rank matrices. Stores sparse plus low-rank matrices
(e.g. from matrix factorization or centering graphs)
of the form \code{x + a \%*\% t(b)} for faster
computation.
}
\section{Slots}{

\describe{
\item{\code{x}}{a sparse matrix}

\item{\code{a}}{a low-rank factor or a matrix}

\item{\code{b}}{optional. a low-rank factor for \code{a \%*\% t(b)}. if \code{b} is not provided, a will be factorized using 
\code{\link[irlba]{irlba}} provided \code{factorize = TRUE}}
}}

\seealso{
Methods are documented in \code{\link{splr}}.
}
