% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splr_matrix_class.R
\name{splr_sparse_plus_constant}
\alias{splr_sparse_plus_constant}
\title{Add a constant to a splrMatrix object}
\usage{
splr_sparse_plus_constant(x, a)
}
\arguments{
\item{x}{splrMatrix object}

\item{a}{scalar}
}
\value{
new splrMatrix object x + a
}
\description{
Add a constant to a splrMatrix object
}
